requirements.txt
autopep8==2.0.1
dnspython==2.3.0
future==0.18.3
pycodestyle==2.10.0
python-whois==0.8.0
tomli==2.0.1
whois==0.9.27
